+++ dist_node/statement.js
@@ -1,4 +1,7 @@
-"use strict";
+/*
+ * THIS FILE IS AUTO GENERATED from 'lib/statement.kep'
+ * DO NOT EDIT
+*/"use strict";
 var __o = require("./node"),
     defineNode = __o["defineNode"],
     Node = __o["Node"],
@@ -95,19 +98,19 @@ var __o = require("./node"),
     (self.update = (update || null));
     (self.body = body);
 })));
-(exports.Statement = Statement);
-(exports.EmptyStatement = EmptyStatement);
-(exports.DebuggerStatement = DebuggerStatement);
-(exports.WithStatement = WithStatement);
-(exports.BlockStatement = BlockStatement);
-(exports.ExpressionStatement = ExpressionStatement);
-(exports.IfStatement = IfStatement);
-(exports.BreakStatement = BreakStatement);
-(exports.ContinueStatement = ContinueStatement);
-(exports.SwitchStatement = SwitchStatement);
-(exports.ReturnStatement = ReturnStatement);
-(exports.ThrowStatement = ThrowStatement);
-(exports.TryStatement = TryStatement);
-(exports.WhileStatement = WhileStatement);
-(exports.DoWhileStatement = DoWhileStatement);
-(exports.ForStatement = ForStatement);
+(exports["Statement"] = Statement);
+(exports["EmptyStatement"] = EmptyStatement);
+(exports["DebuggerStatement"] = DebuggerStatement);
+(exports["WithStatement"] = WithStatement);
+(exports["BlockStatement"] = BlockStatement);
+(exports["ExpressionStatement"] = ExpressionStatement);
+(exports["IfStatement"] = IfStatement);
+(exports["BreakStatement"] = BreakStatement);
+(exports["ContinueStatement"] = ContinueStatement);
+(exports["SwitchStatement"] = SwitchStatement);
+(exports["ReturnStatement"] = ReturnStatement);
+(exports["ThrowStatement"] = ThrowStatement);
+(exports["TryStatement"] = TryStatement);
+(exports["WhileStatement"] = WhileStatement);
+(exports["DoWhileStatement"] = DoWhileStatement);
+(exports["ForStatement"] = ForStatement);